requirements.txt
elevenlabs==1.5.0
python-dotenv==1.0.1
requests==2.31.0
pydantic>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0